chore: type 파일로 분리
